pyproject.toml
[tool.poetry]
name = "remote-hub"
version = "0.2.0"
description = ""
authors = ["Jean Carraro <jgcarraro@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10"
paramiko = "^3.4.0"
pyyaml = "^6.0.1"
scp = "^0.14.5"
questionary = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
transfer = "remote_hub.transfer:main"
ssh_to_server = "remote_hub.ssh_to_server:main"
config_routes = "remote_hub.config_routes:main"
